requirements.txt
fastapi == 0.115.11
motor == 3.7.0
python-dotenv == 1.0.1
uvicorn == 0.34.0